requirements.txt
blobfile>=1.0.5
numpy
Pillow
torch
torchvision
tqdm
